booking confirmations files broken in smaller files
