+++ supabase/functions/send-booking-confirmation/resendClient.ts
@@ -1,0 +1,21 @@
+export async function sendResendEmail({ to, subject, html }: { to: string; subject: string; html: string }) {
+  const RESEND_API_KEY = Deno.env.get('RESEND_API_KEY');
+  if (!RESEND_API_KEY) throw new Error('Missing RESEND_API_KEY');
+
+  const response = await fetch('https://api.resend.com/emails', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${RESEND_API_KEY}`,
+    },
+    body: JSON.stringify({
+      from: 'Stisses Bokning <noreply@svejo.se>',
+      to,
+      subject,
+      html,
+    }),
+  });
+
+  const data = await response.json();
+  return { response, data };
+}
